Fix incorrect text filter on going back in Channels
