+++ src/tribler-gui/tribler_gui/widgets/searchresultswidget.py
@@ -11,9 +11,8 @@ from tribler_common.sentry_reporter.sentry_mixin import AddBreadcrumbOnShowMixin
 from tribler_core.modules.metadata_store.serialization import CHANNEL_TORRENT, COLLECTION_NODE, REGULAR_TORRENT
 
 from tribler_gui.tribler_request_manager import TriblerNetworkRequest
-from tribler_gui.utilities import connect, get_ui_file_path, tr
+from tribler_gui.utilities import connect, get_ui_file_path, to_fts_query, tr
 from tribler_gui.widgets.tablecontentmodel import SearchResultsModel
-from tribler_gui.widgets.triblertablecontrollers import to_fts_query
 
 widget_form, widget_class = uic.loadUiType(get_ui_file_path('search_results.ui'))
